added fallback when NODE_ENV is undefined, moved config.module.ts and config.service.ts
