+++ src/services/config.service.ts
@@ -13,7 +13,9 @@ export class ConfigService {
   private readonly envConfig: { [key: string]: string };
 
   constructor(filePath: string) {
-    this.envConfig = dotenv.parse(fs.readFileSync(filePath));
+    if (process.env.NODE_ENV) {
+      this.envConfig = dotenv.parse(fs.readFileSync(filePath));
+    }
   }
 
   get(key: string): string {
